+++ .gitmodules
@@ -1,0 +1,4 @@
+[submodule "OtterGui"]
+	path = OtterGui
+	url = git@github.com:Ottermandias/OtterGui.git
+	branch = main
