+++ Glamourer/FixedDesigns.cs
@@ -1,0 +1,12 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Glamourer;
+
+public class FixedDesigns
+{
+    public bool TryGetDesign(Actor.IIdentifier actor, [NotNullWhen(true)] out CharacterSave? save)
+    {
+        save = null;
+        return false;
+    }
+}
